@@ -23,18 +23,18 @@
     ],
     "config": {
         "platform": {
-            "php": "7.2.5"
+            "php": "7.4"
         }
     },
     "require": {
-        "php": "^7.1",
+        "php": "^7.4",
         "container-interop/container-interop": "^1.1",
         "guzzlehttp/command": "^1.0",
         "guzzlehttp/guzzle": "^7.0",
         "guzzlehttp/guzzle-services": "^1.0",
         "guzzlehttp/psr7": "^2.0",
         "psr/http-message": "^1.0",
-        "lcobucci/jwt": "^3.2"
+        "lcobucci/jwt": "^4.0"
     },
     "require-dev": {
         "phpunit/phpunit": "^8.0",
